fix: fixes in services

+++ srcs/requirements/nginx/tools/run_nginx.sh
@@ -4,12 +4,14 @@ openssl req -x509 -nodes -days 365 -newkey rsa:2048 -keyout $KEYOUT_ -out $CERTS
 
 cd /etc/nginx/sites-available/
 
-sed -i "s#\$DOMAIN_NAME#$DOMAIN_NAME#g" default.conf
-sed -i "s#\$CERTS_#$CERTS_#g" default.conf
-sed -i "s#\$KEYOUT_#$KEYOUT_#g" default.conf
+if [ -f ./default.conf ]; then
+    sed -i "s#\$DOMAIN_NAME#$DOMAIN_NAME#g" default.conf
+    sed -i "s#\$CERTS_#$CERTS_#g" default.conf
+    sed -i "s#\$KEYOUT_#$KEYOUT_#g" default.conf
 
-cat default.conf > default
+    cat default.conf > default
 
-rm -rf default.conf
+    rm -rf default.conf
+fi
 
 nginx -g "daemon off;"
